add fix for off board temperature & add testing CI pipeline
As per #6 we were returning -1 for out of range temperatures in celcius,
despite that being a very common temperature.

It has been update to return `sys.maxsize` by default,
or to raise a ValueError if requested.

I've not yet fixed the on board temperature, that will follow shortly.

I've added a Github action for running the unit tests using tox.